pyproject.toml
[project]
name = "mamayogabot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp==3.13.2",
    "aiosqlite==0.21.0",
    "alembic==1.17.2",
    "greenlet>=3.3.1",
    "pytelegrambotapi==4.29.1",
    "python-dotenv==1.0.1",
    "sqlalchemy[asyncio]==2.0.44",
]
